App/requirements.txt
fastapi
python-dotenv
uvicorn[standard]
websockets
psycopg2
pandas
jinja2
google-cloud-storage
fastapi_mail
python-multipart
python-jose[cryptography]
uuid
bcrypt
requests
fastapi_limiter
slowapi
sqlalchemy
pydantic
pydantic-settings
psycopg2-binary
asyncpg
pillow
alembic
pytest
pytest-asyncio
httpx
pytest-cov
passlib[bcrypt]
argon2-cffi
aiofiles
boto3
aiohttp
pandas
email_validator
apscheduler
redis
celery
rapidfuzz
openpyxl
websockets
reportlab
PyPDF2
tenacity>=8.2.0
sendgrid
# apscheduler-redis
# git+https://github.com/agronholm/apscheduler_redis_jobstore.git#egg=apscheduler_redis_jobstore
